Move form extension point to extensions.ts + docu
